Remove album from details
The album name is arguably the least important part of the song description, yet in the current implementation it comes before artist. This often causes text clipping in discord, hiding the artist name whilst showing the album it's on.

This also works better for singles, as the name of the song is duplicated in the current implementation (as album name == song name). Overall, it's simply a cleaner way of displaying song info.